pyproject.toml
[project]
name = "rl-chrome-dino"
version = "0.1.0"
description = "RL agent for Chrome Dino using Playwright and Stable-Baselines3"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "gymnasium==0.29.1",
  "numpy==1.26.4",
  "Pillow==10.2.0",
  "playwright==1.42.0",
  "stable-baselines3==2.3.0",
  "setuptools<81",
  "tensorboard==2.16.2",
  "torch==2.2.1",
  "PyYAML==6.0.1",
]

[tool.uv]
package = false
